Add shadow to card, add redirects
