+++ src/Pages/Cart/Cart.css
@@ -1,0 +1,5 @@
+.cart-details{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
