+++ src/styles.css
@@ -44,5 +44,10 @@
 
 .card{
   width: auto;
+  transition: 400ms ease;
+}
+
+.card:hover {
+  box-shadow: 0 0.5rem 1.5rem 0.25rem rgba(0,0,0,0.25);
 }
